Docs: Config / Data improvements
