+++ docs/.vitepress/components/Badge.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 withDefaults(
 	defineProps<{
-		as: string;
+		as?: string;
 	}>(),
 	{
 		as: 'p',
