+++ docs/.vitepress/data/blog.data.ts
@@ -12,7 +12,6 @@ export default defineLoader({
 						{ author: ['first_name', 'last_name', 'avatar', 'title'] },
 						{ tags: [{ directus_tags_id: ['title', 'slug', 'type'] }] },
 					],
-					// @ts-ignore
 					filter: {
 						status: { _eq: 'published' },
 					},
