+++ docs/.vitepress/theme/components/VPSidebarItem.vue
@@ -1,14 +1,14 @@
-<!-- eslint-disable vue/no-v-html vue/no-v-text-v-html-on-component -->
-<script setup lang="ts">
-// Extended version of https://github.com/vuejs/vitepress/blob/main/src/client/theme-default/components/VPSidebarItem.vue
+// Extended version of
+https://github.com/vuejs/vitepress/blob/main/src/client/theme-default/components/VPSidebarItem.vue
 
+<script setup lang="ts">
 import { useData } from 'vitepress';
 // @ts-ignore
 import VPLink from 'vitepress/dist/client/theme-default/components/VPLink.vue';
 // @ts-ignore
 import VPIconChevronRight from 'vitepress/dist/client/theme-default/components/icons/VPIconChevronRight.vue';
 // @ts-ignore
-import { useSidebarControl } from 'vitepress/dist/client/theme-default/composables/sidebar.js';
+import { useSidebarControl } from 'vitepress/dist/client/theme-default/composables/sidebar';
 import type { DefaultTheme } from 'vitepress/theme';
 import { computed, type HTMLAttributes } from 'vue';
