+++ docs/.vitepress/tsconfig.app.json
@@ -1,4 +1,9 @@
 {
 	"extends": "@directus/tsconfig/vue3",
+	"compilerOptions": {
+		"paths": {
+			"@/*": ["./*"]
+		}
+	},
 	"include": ["env.d.ts", "**/*.vue"]
 }
