+++ docs/blog/[slug].md
@@ -7,8 +7,8 @@ type: 'blog-post'
 
 <script setup>
 import { useData } from 'vitepress'
-import Badge from '../.vitepress/components/Badge.vue'
-import PostMeta from '../.vitepress/components/blog/PostMeta.vue'
+import Badge from '@/components/Badge.vue'
+import PostMeta from '@/components/blog/PostMeta.vue'
 const { params } = useData()
 </script>
