+++ docs/package.json
@@ -20,13 +20,15 @@
 		"@directus/format-title": "10.1.0",
 		"@directus/sdk": "workspace:*",
 		"@directus/tsconfig": "workspace:*",
+		"@pnpm/find-workspace-dir": "6.0.2",
+		"@pnpm/find-workspace-packages": "6.0.9",
 		"eslint-parser-plain": "0.1.0",
 		"eslint-plugin-markdown": "3.0.1",
 		"eslint-plugin-prettier": "npm:@paescuj/eslint-plugin-prettier@5.0.1-1",
 		"eslint-plugin-react": "7.33.2",
 		"marked": "10.0.0",
 		"spellchecker-cli": "6.1.1",
-		"typedoc": "0.25.3",
+		"typedoc": "0.25.4",
 		"typedoc-plugin-frontmatter": "0.0.2",
 		"typedoc-plugin-markdown": "4.0.0-next.30",
 		"typedoc-plugin-zod": "1.1.0",
