+++ docs/tsconfig.node.json
@@ -1,7 +1,8 @@
 {
 	"extends": "@directus/tsconfig/node18-esm",
 	"compilerOptions": {
-		"declaration": false
+		"declaration": false,
+		"resolveJsonModule": true
 	},
 	"include": ["**/*", ".vitepress/**/*.ts"]
 }
